@@ -5,34 +5,28 @@
     "build": "webpack --mode=production",
     "analyze": "webpack --mode=production --env analyze",
     "lint": "eslint src",
-    "test": "jest --passWithNoTests",
-    "watch-tests": "jest --watch",
-    "format": "prettier --write .",
-    "coverage": "jest --coverage"
-  },
-  "husky": {
-    "hooks": {
-      "pre-commit": "pretty-quick --staged && concurrently npm run test npm run lint"
-    }
+    "lint:tsc": "tsc --noEmit",
+    "format": "prettier --write ."
   },
   "devDependencies": {
     "@babel/core": "^7.12.10",
     "@babel/plugin-proposal-class-properties": "^7.14.5",
     "@babel/preset-env": "^7.12.11",
     "@babel/preset-react": "^7.14.5",
-    "babel-eslint": "^10.1.0",
-    "babel-loader": "^8.1.0",
-    "concurrently": "^5.2.0",
+    "@babel/preset-typescript": "^7.15.0",
+    "@types/node": "^16.10.9",
+    "@types/webpack": "^5.28.0",
+    "@typescript-eslint/eslint-plugin": "^5.0.0",
+    "@typescript-eslint/parser": "^5.0.0",
     "css-loader": "^5.0.1",
-    "eslint": "^7.18.0",
+    "eslint": "^7.32.0",
     "eslint-config-prettier": "^7.2.0",
-    "eslint-config-react-important-stuff": "^3.0.0",
+    "eslint-config-standard": "^16.0.3",
+    "eslint-plugin-import": "^2.25.2",
     "eslint-plugin-prettier": "^3.3.1",
-    "husky": "^4.3.8",
-    "identity-obj-proxy": "^3.0.0",
-    "lodash": "^4.17.20",
+    "eslint-plugin-react": "^7.26.1",
+    "eslint-plugin-react-hooks": "^4.2.0",
     "prettier": "^2.0.5",
-    "pretty-quick": "^3.1.0",
     "react-router": "^5.2.0",
     "react-router-dom": "^5.2.0",
     "sass": "^1.42.1",
@@ -40,9 +34,11 @@
     "single-spa-react": "^4.0.0",
     "style-loader": "^2.0.0",
     "systemjs-webpack-interop": "^2.3.4",
+    "ts-node": "^10.3.0",
+    "typescript": "^4.4.4",
     "webpack": "^5.18.0",
     "webpack-cli": "^4.4.0",
-    "webpack-config-single-spa-react": "^4.0.2",
+    "webpack-config-single-spa-react-ts": "^4.0.2",
     "webpack-dev-server": "^4.0.0-beta.0",
     "webpack-merge": "^5.7.3"
   },
