@@ -1,14 +1,14 @@
-import React from "react";
-import ReactDOM from "react-dom";
-import singleSpaReact from "single-spa-react";
+import React from 'react';
+import ReactDOM from 'react-dom';
+import singleSpaReact from 'single-spa-react';
 
-import Root from "./root";
+import { Root } from './root';
 
 const lifecycles = singleSpaReact({
   React,
   ReactDOM,
   rootComponent: Root,
-  errorBoundary(err, info, props) {
+  errorBoundary() {
     return <div className="mt-16">Error</div>;
   },
 });
