@@ -1,8 +1,8 @@
-import * as React from "react";
-import ReactDOM from "react-dom";
-import singleSpaReact from "single-spa-react";
+import React from 'react';
+import ReactDOM from 'react-dom';
+import singleSpaReact from 'single-spa-react';
 
-import { Root } from "./root";
+import { Root } from './root';
 
 const lifecycles = singleSpaReact({
   React,
